+++ libs/components/src/lib/breadcrumb/index.ts
@@ -1,14 +1,4 @@
-import type { FoundationElementDefinition } from '@microsoft/fast-foundation';
-import { designSystem, getPrefix } from '../../shared/design-system';
-import { breadcrumbTemplate as template } from './breadcrumb.template';
-import styles from './breadcrumb.scss';
-import { Breadcrumb } from './breadcrumb';
+import { registerBreadcrumb } from './definition';
 
-export const vividBreadcrumb = Breadcrumb.compose<FoundationElementDefinition>({
-	baseName: 'breadcrumb',
-	template,
-	styles,
-});
 
-designSystem.withPrefix(getPrefix(import.meta.url)).register(vividBreadcrumb());
-
+registerBreadcrumb();
